merge v6.0 applications to main

+++ android/test-app-maven-central/src/main/java/com/arthenica/ffmpegkit/test/SubtitleTabFragment.java
@@ -286,7 +286,7 @@ public class SubtitleTabFragment extends Fragment {
             return;
         }
 
-        int timeInMilliseconds = this.statistics.getTime();
+        double timeInMilliseconds = this.statistics.getTime();
         int totalVideoDuration = 9000;
 
         String completePercentage = new BigDecimal(timeInMilliseconds).multiply(new BigDecimal(100)).divide(new BigDecimal(totalVideoDuration), 0, BigDecimal.ROUND_HALF_UP).toString();
