+++ macos/test-app-cocoapods/FFmpegKitMACOS/VideoViewController.m
@@ -285,7 +285,7 @@
         return;
     }
 
-    int timeInMilliseconds = [statistics getTime];
+    double timeInMilliseconds = [statistics getTime];
     int totalVideoDuration = 9000;
 
     int percentage = timeInMilliseconds*100/totalVideoDuration;
